+++ x/compliance/types/message_certify_model.go
@@ -9,7 +9,7 @@ const TypeMsgCertifyModel = "certify_model"
 
 var _ sdk.Msg = &MsgCertifyModel{}
 
-func NewMsgCertifyModel(signer string, vid int32, pid int32, softwareVersion uint64, softwareVersionString string, certificationDate string, certificationType string, reason string) *MsgCertifyModel {
+func NewMsgCertifyModel(signer string, vid int32, pid int32, softwareVersion uint32, softwareVersionString string, certificationDate string, certificationType string, reason string) *MsgCertifyModel {
 	return &MsgCertifyModel{
 		Signer:                signer,
 		Vid:                   vid,
